backend/requirements.txt
# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Deep Learning - PyTorch
torch>=2.0.0
torchvision>=0.15.0

# HuggingFace Transformers for BERT
transformers>=4.30.0
tensorflow>=2.13.0
sentencepiece>=0.1.99

# Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# Machine Learning Utilities
scikit-learn>=1.3.0
scipy>=1.10.0

# Heatmap Visualization (Grad-CAM)
grad-cam>=1.4.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.24.0
